[aa] fix input validator

+++ aa/input_format_validators/validator/validator.cpp
@@ -20,6 +20,6 @@ void run() {
 		words.insert(w);
 	}
 
-	assert(total <= Arg("m"));
+	assert(total <= (int)Arg("m"));
 	assert((int)words.size() == n);
 }
